# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Layblar Platform API
  description: "# Introduction\nThis is the API Documentation for the Layblar-Platform.\
    \ \n"
  contact:
    name: Layblar Documentation
    url: https://layblar.github.io/docs
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
  x-logo:
    url: https://layblar.github.io/docs/img/layblarlogo.svg
servers:
- url: "http://{tenant}"
  variables:
    tenant:
      default: localhost:8080
...
paths:
...
  /auth/register/researcher:
    post:
      tags:
      - Authorization
      summary: Register a new researcher
      description: Creates a new researcher
      operationId: createResearcher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterResearcherDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "201":
          description: The registered researcher
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  researcherId:
                    type: string
                  email:
                    type: string
                  householdId:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
      security:
      - none: []
...
  /household/{householdId}:
    get:
      tags:
      - Household
      summary: Get household information
      description: Get information about the given household
      operationId: getHouseholdInformation
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the data should be fetched
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Household with the given id
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
...
  /project/project:
    get:
      tags:
      - Project
      summary: List all Projects
      description: List all Projects
      operationId: listProjects
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: List of all projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectInfoDTO'
      security:
      - jwt: []
    post:
      tags:
      - Project
      summary: Create a Project
      description: Create a new Project
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The created project
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectName:
                    type: string
                  projectDescription:
                    type: string
                  projectDataUseDeclartion:
                    type: string
                  startDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  endDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  created:
                    $ref: '#/components/schemas/LocalDateTime'
                  metaData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMetaDataDTO'
                  labels:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabelDTO'
      security:
      - jwt: []
  /project/project/{projectId}:
    get:
      tags:
      - Project
      summary: Get a Project
      description: Get a Project
      operationId: getProject
      parameters:
      - name: projectId
        in: path
        description: The ID of the project to get
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Project by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectName:
                    type: string
                  projectDescription:
                    type: string
                  projectDataUseDeclartion:
                    type: string
                  startDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  endDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  created:
                    $ref: '#/components/schemas/LocalDateTime'
                  metaData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMetaDataDTO'
                  labels:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabelDTO'
      security:
      - jwt: []
    put:
      tags:
      - Project
      summary: Update a Project
      description: Update a Project
      operationId: updateProject
      parameters:
      - name: projectId
        in: path
        description: The ID of the project to update
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The updated project
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectName:
                    type: string
                  projectDescription:
                    type: string
                  projectDataUseDeclartion:
                    type: string
                  startDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  endDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  created:
                    $ref: '#/components/schemas/LocalDateTime'
                  metaData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMetaDataDTO'
                  labels:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabelDTO'
      security:
      - jwt: []
  /project/project/{projectId}/data:
    get:
      tags:
      - Project
      summary: Get data from a Project
      description: Get data from a Project
      operationId: getProjectData
      parameters:
      - name: projectId
        in: path
        description: The ID of the project to get data from
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Project data by id
          content:
            application/json:
              schema:
                type: object
      security:
      - jwt: []
  /project/project/{projectId}/household/{householdId}:
    post:
      tags:
      - Project
      summary: Join a Project
      description: Join a Project
      operationId: joinProject
      parameters:
      - name: householdId
        in: path
        description: The ID of the household that wants to join the project
        required: true
        schema:
          type: string
      - name: projectId
        in: path
        description: The ID of the project to join
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinProjectDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The joined Project Information
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectName:
                    type: string
                  projectDescription:
                    type: string
                  projectDataUseDeclartion:
                    type: string
                  startDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  endDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  created:
                    $ref: '#/components/schemas/LocalDateTime'
                  metaData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMetaDataDTO'
                  labels:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabelDTO'
      security:
      - jwt: []
  /smartmeter/household/{householdId}:
    get:
      tags:
      - Smart-Meter
      summary: Get Smart-Meter data
      description: Get a List of Smart-Meter data from the specified time frame
      operationId: getSmartMeterData
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the data should be fetched
        required: true
        schema:
          type: string
      - name: from
        in: query
        description: The begin of the time intervall for which the data should be
          fetched. Needs to be a Unix Timestamp as String
        required: true
        schema:
          type: string
      - name: to
        in: query
        description: The end of the time intervall for which the data should be fetched.
          Needs to be a Unix Timestamp as String
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "400":
          description: Missing Query Parameters
        "422":
          description: Wrong Date Format
        "200":
          description: Smart-Meter data for the specified time period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SmartMeterDataDTO'
      security:
      - jwt: []
components:
  schemas:
...
    CreateProjectDTO:
      type: object
      properties:
        projectName:
          type: string
        projectDescription:
          type: string
        projectDataUseDeclartion:
          type: string
        startDate:
          $ref: '#/components/schemas/LocalDateTime'
        endDate:
          $ref: '#/components/schemas/LocalDateTime'
        metaData:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMetaDataDTO'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelDTO'
    DeviceCategoryDTO:
      type: object
      properties:
        deviceCategoryId:
          type: string
        deviceCategoryName:
          type: string
        deviceCategoryDescription:
          type: string
    DeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
    HouseholdDeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
    HouseholdUserDTO:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    JoinProjectDTO:
      type: object
      properties:
        householdId:
          type: string
        metaData:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMetaDataDTO'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceDTO'
    LabelDTO:
      type: object
      properties:
        labelId:
          type: string
        labelName:
          type: string
        labelDescription:
          type: string
        labelMethod:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
...
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
    ProjectInfoDTO:
      type: object
      properties:
        projectId:
          type: string
        projectName:
          type: string
        projectDescription:
          type: string
        projectDataUseDeclartion:
          type: string
        startDate:
          $ref: '#/components/schemas/LocalDateTime'
        endDate:
          $ref: '#/components/schemas/LocalDateTime'
        created:
          $ref: '#/components/schemas/LocalDateTime'
        metaData:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMetaDataDTO'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelDTO'
    ProjectMetaDataDTO:
      type: object
      properties:
        metaDataId:
          type: string
        metaDataName:
          type: string
        type:
          type: string
        isRequired:
          type: boolean
...
    RegisterResearcherDTO:
      type: object
      properties:
        password:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        organization:
          type: string
    SmartMeterDTO:
      type: object
      properties:
        smartMeterId:
          type: string
    SmartMeterDataDTO:
      description: Smart Meter Data
      type: object
      properties:
        time:
          description: Time at which Smart-Meter data was recored
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          example: 2023-05-06T11:15:40
        sensorId:
          description: Id of the Smart-Meter that recored the meassurement
          type: string
          example: 123MF34FBDKAS
        "1.7.0":
          format: float
          description: This is the current positive active power in watts measured
            from Phase A.
          type: number
        "1.8.0":
          format: float
          description: This is the total positive active energy in watt-hours consumed
            or generated from Phase A.
          type: number
        "2.7.0":
          format: float
          description: This is the current negative active power in watts measured
            from Phase A. This could indicate generated energy.
          type: number
        "2.8.0":
          format: float
          description: This is the total negative active energy in watt-hours consumed
            or generated from Phase A.
          type: number
        "3.8.0":
          format: float
          description: This is the total positive reactive energy in var-hours consumed
            or generated from Phase A.
          type: number
        "4.8.0":
          format: float
          description: This is the total negative reactive energy in var-hours consumed
            or generated from Phase A
          type: number
        "16.7.0":
          format: float
          description: This is the difference between positive and negative active
            instantaneous power in watts from Phase A.
          type: number
        "31.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L1.
          type: number
        "32.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L1.
          type: number
        "51.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L2.
          type: number
        "52.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L2.
          type: number
        "71.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L3.
          type: number
        "72.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L3.
          type: number
...
    UpdateProjectDTO:
      type: object
      properties:
        projectId:
          type: string
        projectName:
          type: string
        projectDescription:
          type: string
        projectDataUseDeclartion:
          type: string
        startDate:
          $ref: '#/components/schemas/LocalDateTime'
        endDate:
          $ref: '#/components/schemas/LocalDateTime'
        metaData:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMetaDataDTO'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelDTO'
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    jwt:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    none:
      type: http
      description: No Authentication
      scheme: No-Auth