# projectservice/src/main/java/at/fhv/layblar/rest/ProjectResource.java
package at.fhv.layblar.rest;

import at.fhv.layblar.application.ProjectService;
import at.fhv.layblar.commands.CreateProjectCommand;
import at.fhv.layblar.commands.JoinProjectCommand;
import at.fhv.layblar.commands.RegisterResearcherCommand;
import at.fhv.layblar.commands.UpdateProjectCommand;
import at.fhv.layblar.utils.ResponseExceptionBuilder;
import at.fhv.layblar.utils.exceptions.ResponseException;
import io.quarkus.security.Authenticated;
import jakarta.inject.Inject;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.PUT;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.core.Response;

@Path("/api")
@Authenticated
public class ProjectResource {

    @Inject
    ProjectService projectService;

    @POST
    @Path("/researcher")
    public Response registerResearcher(RegisterResearcherCommand command){
        return Response.ok().entity(projectService.createResearcher(command)).build();
    }

    @POST
    @Path("/project")
    public Response createProject(CreateProjectCommand command){
        try {
            return Response.ok().entity(projectService.createProject(command)).build();
        } catch (ResponseException e) {
            return ResponseExceptionBuilder.buildResponse(e);
        }
    }

    @PUT
    @Path("/project/{projectId}")
    public Response updateProject(@PathParam("projectId") String projectId, UpdateProjectCommand command){
        try {
            return Response.ok().entity(projectService.updateProject(projectId, command)).build();
        } catch (ResponseException e) {
            return ResponseExceptionBuilder.buildResponse(e);
        }
    }

    @GET
    @Path("/project/{projectId}")
    public Response getProject(@PathParam("projectId") String projectId){
        try {
            return Response.ok().entity(projectService.getProjectInfo(projectId)).build();
        } catch (ResponseException e) {
            return ResponseExceptionBuilder.buildResponse(e);
        }
    }

    @GET
    @Path("/project/{projectId}/data")
    public Response getProjectData(@PathParam("projectId") String projectId){
        return Response.ok().entity(projectService.getProjectData(projectId)).build();
    }

    @GET
    @Path("/project")
    public Response listProjects(){
        return Response.ok().entity(projectService.getProjects()).build();
    }

    @POST
    @Path("/project/{projectId}/household/{householdId}")
    public Response joinProject(@PathParam("projectId") String projectId, @PathParam("householdId") String householdId, JoinProjectCommand command){
        try {
            return Response.ok().entity(projectService.joinProject(projectId, householdId, command)).build();
        } catch (ResponseException e) {
            return ResponseExceptionBuilder.buildResponse(e);
        }
    }

}